Issue not detected with plegal and mlegal
The change is necessary specificly when the nombre_padre or nombre_madre
are in social form or has honorifics, in those cases the condition
nombre_<>.startswith() is not enough and gives a bad result.
The issue is addressed adding two conditions nombre_<>.endswith
(for the social form) and madre in nombre_madre (when there exists honorifics).
It improves the results with the test_data.