Exact name match dataframes to EdgeList format
I concatenate exact name match dataframes of padres and
madres in a unique two column (cedula, matched_cedula)
dataframe which is converted to txt file with spaces as
separation.

This output is the EdgeList format which is used then in
the graph/prototype_graph_0.py script which creates the
graph.