Match by cedula correction
Filter records with cedula before the implementation of the
exact_name_padre() and exact_name_madre() function. This
would reduce the running time of the exact_name() algorithm
since it will review less records.

However, this avoid the comparison between the official
cedula (only empty records with four part names are reviewed)
and the matched cedula.