match.py algorithm corrected
I will list the changes to every function:

match_by_cedula_padre() and match_by_cedula_madre():
Correct the filter of records without cedulas. This way we
take advantage of the data available for the matching.

exact_name() function was deleted and now the functions
exact_name_madre() and exact_name_padre() are directly
called on make_dataset.py.
Also they now filter the records that have cedulas of
padre/madre and the columns padre/madre_official were deleted.

ceds_found() now have the only_cedula option when we are only
interested in cedulas with padre or madre cedula and not the
matching.