Plot Birth Year Largest Subgraph and Node Degree 0
The code describe the manipulation of the network and the data
for the visualization of the birth year distribution of the
largest connected component and of the nodes of degree 0 (do
not appear on the network)

The commit address the issue using mainly two networkit's commands:
1. nk.components.ConnectedComponents.extractLargestConnectedComponent
to extract the largest connected component
2. nk.graphio.EdgeListReader().getNodeMap() to know the relation
between cedulas and the networkit's node ids (automatically created).

The code can be easily adapted to the analysis of other network problems.