Generate centrality measures for each firm
firms_analysis.py contains the workflow of obtaining
centrality measures. We have some important parameters:

- separation_map_owners: generate consanguinity map for owners
- separation_map_employees: generate consanguinity map for employees
- centrality: Obtain centrality measures for each firm: harmonic
centraliy, avg distance employees to onwer, fraction of employees
that are family, fraction of family that are employees and fraction
of citizens that is family.
- avg_closeness_employees: generate average harmonic centrality of
the employees within the firm members.
- generate_output: generate csv.

utils_firms_analysis.py contains all the functions used in
firm_analysis.py